requirements.txt
langchain
langchain-community
openai
faiss-cpu
gradio
PyMuPDF
python-docx
 sentence-transformers 
 faiss-cpu 
requests
